refactor: move calendar to client component
